dialog refactor and base fragment implementation
